+++ sso/src/controllers/dashboard/users/list.ts
@@ -1,9 +1,12 @@
 import { Context, Middleware } from 'koa'
 import { client } from '../../../models/db'
+import { itinerarySlugSchema } from '../../../schemas/itineraries/itinerarySchema'
+import { userStatusSchema } from '../../../schemas/users/userSchema'
 
 export const dashboardListUsers: Middleware = async (ctx: Context) => {
-  const queryResult = await client.query(
-    `SELECT
+  const { itinerarySlug, status } = ctx.state.query
+  let query = `
+  SELECT
     u.id,
     u.name AS name,
     i.name AS "itineraryName",
@@ -11,8 +14,24 @@ export const dashboardListUsers: Middleware = async (ctx: Context) => {
     TO_CHAR(u.created_at, 'YYYY-MM-DD') AS "createdAt"
   FROM
     "user" u
-  JOIN itinerary i ON u.itinerary_id = i.id;`
-  )
+  JOIN itinerary i ON u.itinerary_id = i.id
+`
+  const queryParams = []
+  const conditions = []
+  if (itinerarySlug) {
+    const parsedSlug = itinerarySlugSchema.parse(itinerarySlug)
+    conditions.push(`i.slug = $${conditions.length + 1}`)
+    queryParams.push(parsedSlug)
+  }
+  if (status) {
+    const parsedStatus = userStatusSchema.parse(status)
+    conditions.push(`u.status = $${conditions.length + 1}`)
+    queryParams.push(parsedStatus)
+  }
+  if (conditions.length > 0) {
+    query += ` WHERE ${conditions.join(' AND ')}`
+  }
+  const queryResult = await client.query(query, queryParams)
 
   if (!queryResult.rowCount) {
     ctx.status = 200