+++ sso_front/src/components/layout/index.ts
@@ -1,3 +1,1 @@
-import { Layout } from './Layout'
-
-export { Layout }
+export { Layout } from './Layout'