+++ packages/ui/CHANGELOG.md
@@ -2,11 +2,24 @@
 
 All notable changes to this project will be documented in this file.
 
-## [0.26.0] - 2024-02-19
+## [0.28.0] - 2024-02-19
 
 ### Chore
 
 - Refactor Icon props
+
+## [0.27.0] - 2024-02-21
+
+### Added
+
+- Migrate VoteCounter to UI
+
+### Added
+
+## [0.26.0] - 2024-02-20
+
+- Migrate Tooltip to UI
+- Added target element, hover behavior and position
 
 ## [0.25.0] - 2024-02-19