+++ sso/CHANGELOG.md
@@ -2,6 +2,22 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.13.0] - 2024-02-20
+
+## Added
+
+- Dynamic query filtering for `GET /dashboard/users` by `status`
+
+## Changed
+
+- `createdAt` at `dashboardUsersListSchema` to `z.date()`
+
+## [1.12.0] - 2024-02-19
+
+## Added
+
+- Dynamic query filtering for `GET /dashboard/users` by `itinerarySlug`
+
 ## [1.11.0] - 2024-02-15
 
 ## Added